requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
discord.py==1.7.3
idna==2.10
multidict==5.1.0
pycparser==2.20
python-dotenv==0.17.1
requests==2.25.1
schedule==1.1.0
six==1.16.0
spotipy==2.18.0
typing-extensions==3.10.0.0
urllib3==1.26.5
yarl==1.6.3
